# docs/12-PLUGINS/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Dowze — API Plateforme de Plugins
  version: "1.0.0"
  description: >
    Contrat public `/v1/` de la plateforme de plugins Dowze (P1 : registre & scopes).
    **Politique de dépréciation N/N-1** : une nouvelle version majeure (`/v2/`) ne retire une
    version précédente qu'après une période de chevauchement documentée ; `/v1/` reste servie
    tant que `/v2/` n'est pas stable. Les plugins déclarent l'`apiVersion` qu'ils ciblent dans
    leur manifest. Auth : `Authorization: Bearer <JWT Supabase>` sur toutes les routes.
servers:
  - url: https://api.dowze.ch
security:
  - bearerAuth: []
...
paths:
...
  /v1/ai/compose:
    post:
      tags: [ai]
      summary: "Contexte plugin → prompt lisible (scope ai:infer, déterministe, sans coût)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profileId, sourceApp, title]
              properties:
                profileId: { type: string, format: uuid }
                sourceApp: { type: string }
                title: { type: string }
                goal: { type: string }
                level: { type: string }
                context:
                  type: array
                  items:
                    type: object
                    properties: { label: { type: string }, value: { type: string } }
                instructions: { type: string }
      responses:
        "201":
          description: Prompt + closingPrompt
          content:
            application/json:
              schema:
                type: object
                properties: { prompt: { type: string }, closingPrompt: { type: string } }
        "403": { description: Scope ai:infer non accordé }
  /v1/ai/ingest:
    post:
      tags: [ai]
      summary: "Résumé texte + champs → snapshot structuré (scope ai:infer, crédits/BYOK réutilisés)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profileId, sourceApp, summary, fields]
              properties:
                profileId: { type: string, format: uuid }
                sourceApp: { type: string }
                summary: { type: string }
                fields:
                  type: array
                  items:
                    type: object
                    required: [key, type]
                    properties:
                      key: { type: string }
                      type: { type: string, enum: [string, number, integer, boolean, stringArray] }
                      description: { type: string }
                      optional: { type: boolean }
                modelId: { type: string }
      responses:
        "201":
          description: Snapshot structuré + crédits dépensés
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshot: { type: object, additionalProperties: true }
                  creditsSpent: { type: number }
        "402": { description: Crédits insuffisants (mode crédits) }
        "403": { description: Scope ai:infer non accordé }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...